# Make next route configuration thread local