--- a/src/NSubstitute/Core/ICallRouter.cs
+++ b/src/NSubstitute/Core/ICallRouter.cs
@@ -17,6 +17,8 @@ namespace NSubstitute.Core
         ConfiguredCall LastCallShouldReturn(IReturn returnValue, MatchArgs matchArgs, PendingSpecificationInfo pendingSpecInfo);
         object Route(ICall call);
         IEnumerable<ICall> ReceivedCalls();
+        [Obsolete("This method is deprecated and will be removed in future versions of the product. " +
+                  "Please use " + nameof(IThreadLocalContext) + "." + nameof(IThreadLocalContext.SetNextRoute) + " method instead.")]
         void SetRoute(Func<ISubstituteState, IRoute> getRoute);
         void SetReturnForType(Type type, IReturn returnValue);
         void RegisterCustomCallHandlerFactory(CallHandlerFactory factory);