--- a/src/NSubstitute/Core/ICallRouterProvider.cs
+++ b/src/NSubstitute/Core/ICallRouterProvider.cs
@@ -1,0 +1,7 @@
+namespace NSubstitute.Core
+{
+    public interface ICallRouterProvider
+    {
+        ICallRouter GetCallRouter();
+    }
+}