--- a/src/NSubstitute/Core/IThreadLocalContext.cs
+++ b/src/NSubstitute/Core/IThreadLocalContext.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using NSubstitute.Core.Arguments;
+using NSubstitute.Routing;
 
 namespace NSubstitute.Core
 {
@@ -11,6 +12,16 @@ namespace NSubstitute.Core
         void SetLastCallRouter(ICallRouter callRouter);
         void ClearLastCallRouter();
         ConfiguredCall LastCallShouldReturn(IReturn value, MatchArgs matchArgs);
+
+        /// <summary>
+        /// Sets the route to use for the next call dispatch on the current thread for the specified <paramref name="callRouter"/>.
+        /// </summary>
+        void SetNextRoute(ICallRouter callRouter, Func<ISubstituteState, IRoute> nextRouteFactory);
+        /// <summary>
+        /// Returns the previously configured next route and resets the stored value.
+        /// If route was configured for the different router, returns <see langword="null"/> and persist the route info.
+        /// </summary>
+        Func<ISubstituteState, IRoute> UseNextRoute(ICallRouter callRouter);
 
         void EnqueueArgumentSpecification(IArgumentSpecification spec);
         IList<IArgumentSpecification> DequeueAllArgumentSpecifications();