--- a/src/NSubstitute/Routing/IRoute.cs
+++ b/src/NSubstitute/Routing/IRoute.cs
@@ -5,6 +5,5 @@ namespace NSubstitute.Routing
     public interface IRoute
     {
         object Handle(ICall call);
-        bool IsRecordReplayRoute { get; }
     }
 }