Closes #453

I reviewed the issue 453 for a few times and while personally I don't want to support concurrency on setup/teardown phases, I feel that NSubstitute should be transparent on the behavior it has. The next route we are specifying for the call router should be thread-local, as our syntax in fact implies that.

After a couple of days of brainstorming, I came up with the current implementation. Now background calls don't affect the arrange/assertion phases of the test and temporary next route is local to the current thread only. I don't believe that someone will really utilize this feature, but I am happy that now we better comply with the [Principle of least astonishment](https://en.wikipedia.org/wiki/Principle_of_least_astonishment) - when people see `subs.Received().Call()` they could barely imagine why background thread becomes crazy 😅 

Notice, I needed to apply a change to our internal architecture. Previously, the `CallRouter` instance was mixed in to the substitute directly (meaning that substitute implemented `ICallRouter` interface). I don't perform that anymore and now I mix in the `CallRouterProvider` wrapper holding a `CallRouter` reference. That's because when I resolve `callRouter` instance from substitute, I need it to point to the actual call router object, rather than to a proxy. See more detail in the code.

Just in case attaching the performance tests:

BEFORE
```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-6700K CPU 4.00GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=3914063 Hz, Resolution=255.4890 ns, Timer=TSC
.NET Core SDK=2.1.402
  [Host] : .NET Core 1.1.2 (CoreCLR 4.6.25211.01, CoreFX 4.6.24705.01), 64bit RyuJIT
  Clr    : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3190.0
  Core   : .NET Core 1.1.2 (CoreCLR 4.6.25211.01, CoreFX 4.6.24705.01), 64bit RyuJIT

                           Method |  Job | Runtime |      Mean |     Error |    StdDev |
--------------------------------- |----- |-------- |----------:|----------:|----------:|
        CreateInterfaceSubstitute |  Clr |     Clr |  5.333 us | 0.0170 us | 0.0159 us |
    CreateAbstractClassSubstitute |  Clr |     Clr |  4.792 us | 0.0152 us | 0.0135 us |
 CreateNonAbstractClassSubstitute |  Clr |     Clr |  4.830 us | 0.0143 us | 0.0127 us |
         CreateDelegateSubstitute |  Clr |     Clr | 10.941 us | 0.0444 us | 0.0416 us |
        CreateInterfaceSubstitute | Core |    Core |  5.581 us | 0.0204 us | 0.0181 us |
    CreateAbstractClassSubstitute | Core |    Core |  5.312 us | 0.0210 us | 0.0186 us |
 CreateNonAbstractClassSubstitute | Core |    Core |  5.221 us | 0.0207 us | 0.0162 us |
         CreateDelegateSubstitute | Core |    Core | 11.754 us | 0.0378 us | 0.0335 us |
```

AFTER:
```
                           Method |  Job | Runtime |      Mean |     Error |    StdDev |
--------------------------------- |----- |-------- |----------:|----------:|----------:|
        CreateInterfaceSubstitute |  Clr |     Clr |  4.552 us | 0.0567 us | 0.0530 us |
    CreateAbstractClassSubstitute |  Clr |     Clr |  4.557 us | 0.0620 us | 0.0518 us |
 CreateNonAbstractClassSubstitute |  Clr |     Clr |  4.377 us | 0.0518 us | 0.0484 us |
         CreateDelegateSubstitute |  Clr |     Clr |  9.769 us | 0.0870 us | 0.0814 us |
        CreateInterfaceSubstitute | Core |    Core |  5.585 us | 0.0655 us | 0.0612 us |
    CreateAbstractClassSubstitute | Core |    Core |  5.021 us | 0.0951 us | 0.0934 us |
 CreateNonAbstractClassSubstitute | Core |    Core |  5.330 us | 0.0690 us | 0.0646 us |
         CreateDelegateSubstitute | Core |    Core | 11.270 us | 0.1117 us | 0.1045 us |
```

It seems that it helped to activate substitute even a bit quicker 😉 

@dtchepak @alexandrnikitin Happy review!