@dtchepak 2018-11-01
@tenninebt It should be any day now, it's mainly pending some doc changes before the release.