@zvirja 2019-03-17
@dtchepak After the change [this doc article](https://nsubstitute.github.io/help/callbacks/) looks a bit irrelevant. While it's still a good practice to not do it, it seems we fully support the scenario now.

What would you suggest to do with the article? We might e.g. add a disclaimer that since v4 it's no longer an issue, but it's still recommended to not do it, as behavior on background threads is subject to race conditions... But it's more a recommendation - not the hard rule as it was before.

Or do you see scenarios when it still might be an issue?