@dtchepak 2019-03-17
@zvirja That link goes to the callback page? Did you mean the [threading](https://nsubstitute.github.io/help/threading/) page?