@zvirja 2019-03-17
@dtchepak I like the text you suggested. But now it's questionable whether this information deserves a whole page, as actually it's not so critical. Should we probably put it to the [FAQ page](https://github.com/nsubstitute/NSubstitute/issues/502)? The section might be named "Thread safety". So we describe that we expect NSubstitute be fully thread-safe, but people better not configure NSubstitute while used in other thread to not get unpredictable results...

But it's just an advise. If you believe a separate page is better - keep it 😉